feat(auth): Add meteorAuthMiddleware for setting userId on the request (#187)
* Add meteorAuthMiddleware for setting userId on the request

This is a more standard approach then detecting userId somewhere inside
the GraphQL middleware. This also uses the standard `Authorization`
header instead of a custom Meteor token header.

* Test if authentication actually works

for both DDP and HTTP requests

* Fix meteorAuthMiddleware not finding the token

* Rename request handler to middleware

because that’s the standard naming of connect and express “plugins”

* Rename bearer to tokenType

* Rename getUserByLoginToken to getUserIdByLoginToken

This makes sure we only get the userId and no other details from the user. That is something for developers to decide and more in line with DDP itself.